+++ wims/src/Misc/symtext/compile.c
@@ -289,8 +289,8 @@ void cp_cutline(char *p, struct block *blk, int next)
     int l, ll, n, idx, start, end;
     struct block *b;
 
-    if(debug>=3) fprintf(stderr,"Cutline %ld/%d for %.15s.\n",
-			 blk-blockbuf, nextblock, p);
+    if(debug>=3) fprintf(stderr,"Cutline %d/%d for %.15s.\n",
+			 (int)(blk-blockbuf), nextblock, p);
     for(p1=strstr(p,"[|]"); p1; p1=strstr(p1,"[|]")) memmove(p1," | ",3);
     macro_trans(p);
     singlespace(p);
@@ -318,8 +318,8 @@ void cp_cutline(char *p, struct block *blk, int next)
 	    memmove(buf,p1,ll); buf[ll]=0;
 	    if(idx==start) b=blk; else b=blockbuf+idx-1;
 	    if(idx<end) n=idx; else n=next;
-	    if(debug>=3) fprintf(stderr,"String block %ld/%d for %.15s.\n",
-			 b-blockbuf, nextblock, buf);
+	    if(debug>=3) fprintf(stderr,"String block %d/%d for %.15s.\n",
+			 (int)(b-blockbuf), nextblock, buf);
 	    cp_string(buf,b,n);
 	    idx++;
 	}
@@ -687,8 +687,8 @@ void cp_oneblock(char *p, struct block *blk, int next)
 {
     char *pp, *pe;
     int i;
-    if(debug>=3) fprintf(stderr,"Oneblock %ld/%d for %.15s.\n",
-			 blk-blockbuf, nextblock, p);
+    if(debug>=3) fprintf(stderr,"Oneblock %d/%d for %.15s.\n",
+			 (int)(blk-blockbuf), nextblock, p);
     if(myisupper(*p)) {
 	for(pe=p; pe-p < MAX_BINAME && myisalpha(*pe); pe++);
 	if(*pe==':') {
